tools/se_open_data/rubyfiles/triplestore.rb - don't delete the file if it doesn't exist
Attempting to raises an error.

+++ rubyfiles/triplestore.rb
@@ -70,6 +70,6 @@ HERE
   end
 rescue
   # Delete this output file, and rethrow
-  File.delete config.VIRTUOSO_SCRIPT_LOCAL
+  File.delete config.VIRTUOSO_SCRIPT_LOCAL if File.exist? config.VIRTUOSO_SCRIPT_LOCAL
   raise
 end
